+++ src/Commands/SlackChannelsUpdate.php
@@ -14,9 +14,9 @@ use Seat\Slackbot\Exceptions\SlackSettingException;
 use Seat\Slackbot\Helpers\SlackApi;
 use Seat\Slackbot\Models\SlackChannel;
 
-class SlackUpdateChannels extends Command
+class SlackChannelsUpdate extends Command
 {
-    protected $signature = 'slack:update:channels';
+    protected $signature = 'slack:channels:update';
 
     protected $description = 'Discovering Slack channels (both public and private)';
 
@@ -32,35 +32,40 @@ class SlackUpdateChannels extends Command
         if ($token == null) {
             throw new SlackSettingException("missing slack_token in settings");
         }
-        
+
         $api = new SlackApi($token);
 
         $channels = array_merge($api->channels(false), $api->channels(true));
 
         foreach ($channels as $channel) {
             $slackChannel = SlackChannel::find($channel['id']);
+            $isGroup = true;
+            $isGeneral = false;
+
+            // Determine if this is a group (private channel) or a channel
+            if (substr($channel['id'], 0, 1) === 'C') {
+                $isGroup = false;
+            }
+
+            if ($isGroup == false) {
+                $isGeneral = (boolean) $channel['is_general'];
+            }
 
             if ($slackChannel == null) {
-                $slackChannel = new SlackChannel();
-                $slackChannel->id = $channel['id'];
-                $slackChannel->name = $channel['name'];
-                // set private channel flag to true by default
-                $slackChannel->is_group = true;
-
-                // Determine if this is a group (private channel) or a channel
-                if (substr($channel['id'], 0, 1) === 'C') {
-                    $slackChannel->is_group = false;
-                }
-
-                $slackChannel->save();
+                SlackChannel::create([
+                    'id' => $channel['id'],
+                    'name' => $channel['name'],
+                    'is_group' => $isGroup,
+                    'is_general' => $isGeneral
+                ]);
 
                 continue;
             }
 
             $slackChannel->update([
-                'name' => $channel['name']
+                'name' => $channel['name'],
+                'is_general' => $isGeneral
             ]);
         }
-
     }
 }