+++ src/Jobs/AbstractSlack.php
@@ -16,6 +16,7 @@ use Seat\Services\Settings\Seat;
 use Seat\Slackbot\Exceptions\SlackSettingException;
 use Seat\Slackbot\Helpers\SlackApi;
 use Seat\Slackbot\Models\SlackChannelPublic;
+use Seat\Slackbot\Models\SlackLog;
 use Seat\Slackbot\Models\SlackUser;
 use Seat\Web\Models\User;
 
@@ -138,6 +139,7 @@ abstract class AbstractSlack
             ->select('channel_id')
             ->where('users.id', $slackUser->user_id)
             ->where('slack_channels.is_group', (int) $private)
+            ->where('slack_channels.is_general', (int) false)
             ->union(
                 // fix model declaration calling the table directly
                 DB::table('role_user')->join('slack_channel_roles', 'slack_channel_roles.role_id', '=',
@@ -145,6 +147,7 @@ abstract class AbstractSlack
                     ->join('slack_channels', 'slack_channel_roles.channel_id', '=', 'slack_channels.id')
                     ->where('role_user.user_id', $slackUser->user_id)
                     ->where('slack_channels.is_group', (int) $private)
+                    ->where('slack_channels.is_general', (int) false)
                     ->select('channel_id')
             )->union(
                 ApiKey::join('account_api_key_info_characters', 'account_api_key_info_characters.keyID', '=',
@@ -154,6 +157,7 @@ abstract class AbstractSlack
                     ->join('slack_channels', 'slack_channel_corporations.channel_id', '=', 'slack_channels.id')
                     ->where('eve_api_keys.user_id', $slackUser->user_id)
                     ->where('slack_channels.is_group', (int) $private)
+                    ->where('slack_channels.is_general', (int) false)
                     ->select('channel_id')
             )->union(
                 CharacterSheet::join('slack_channel_alliances', 'slack_channel_alliances.alliance_id', '=',
@@ -164,10 +168,12 @@ abstract class AbstractSlack
                     ->join('eve_api_keys', 'eve_api_keys.key_id', '=', 'account_api_key_info_characters.keyID')
                     ->where('eve_api_keys.user_id', $slackUser->user_id)
                     ->where('slack_channels.is_group', (int) $private)
+                    ->where('slack_channels.is_general', (int) false)
                     ->select('channel_id')
             )->union(
                 SlackChannelPublic::join('slack_channels', 'slack_channel_public.channel_id', '=', 'slack_channels.id')
                     ->where('slack_channels.is_group', (int) $private)
+                    ->where('slack_channels.is_general', (int) false)
                     ->select('channel_id')
             )->get();
 
@@ -176,5 +182,36 @@ abstract class AbstractSlack
         }
 
         return $channels;
+    }
+
+    protected function logEvent($eventType, $channels = null)
+    {
+        $message = '';
+        $channelsString = '';
+
+        if ($channels != null) {
+            $channelsString = implode(',', $channels);
+        }
+
+        switch ($eventType)
+        {
+            case 'invite':
+                $message = 'The user ' . $this->user->name . ' has been invited to following channels : ' .
+                    $channelsString;
+                break;
+            case 'kick':
+                $message = 'The user ' . $this->user->name . ' has been kicked from following channels : ' .
+                    $channelsString;
+                break;
+            case 'mail':
+                $message = 'The mail address for user ' . $this->user->name . ' has not been set (' .
+                    $this->user->email . ')';
+                break;
+        }
+
+        SlackLog::create([
+            'event' => $eventType,
+            'message' => $message
+        ]);
     }
 }