+++ src/Jobs/SlackAssKicker.php
@@ -8,7 +8,6 @@
 namespace Seat\Slackbot\Jobs;
 
 use Seat\Eveapi\Models\Eve\ApiKey;
-use Seat\Slackbot\Models\SlackLog;
 use Seat\Slackbot\Models\SlackUser;
 
 class SlackAssKicker extends AbstractSlack
@@ -34,21 +33,16 @@ class SlackAssKicker extends AbstractSlack
             // if key are not valid OR account no longer paid
             // kick the user from all channels to which he's member
             if ($this->isEnabledKey($keys) == false || $this->isActive($keys) == false) {
-                $this->processChannelsKick($slackUser, $channels);
 
-                $slackLog = new SlackLog();
-                $slackLog->event = 'kick';
-                $slackLog->message = 'The user ' . $this->user->name .
-                    ' has been kicked from following channels : ' . implode(',', $channels);
-                $slackLog->save();
+                if (!empty($channels)) {
+                    $this->processChannelsKick($slackUser, $channels);
+                    $this->logEvent('kick', $channels);
+                }
 
-                $this->processGroupsKick($slackUser, $groups);
-
-                $slackLog = new SlackLog();
-                $slackLog->event = 'kick';
-                $slackLog->message = 'The user ' . $this->user->name .
-                    ' has been kicked from following channels : ' . implode(',', $groups);
-                $slackLog->save();
+                if (!empty($groups)) {
+                    $this->processGroupsKick($slackUser, $groups);
+                    $this->logEvent('kick', $groups);
+                }
 
                 return;
             }
@@ -56,25 +50,21 @@ class SlackAssKicker extends AbstractSlack
             // in other way, compute the gap and kick only the user
             // to channel from which he's no longer granted to be in
             $allowedChannels = $this->allowedChannels($slackUser, false);
-            $allowedGroups = $this->allowedChannels($slackUser, true);
+            $extraChannels = array_diff($channels, $allowedChannels);
 
             // remove channels in which user is already in from all granted channels and invite him
-            $this->processChannelsKick($slackUser, array_diff($channels, $allowedChannels));
-
-            $slackLog = new SlackLog();
-            $slackLog->event = 'kick';
-            $slackLog->message = 'The user ' . $this->user->name .
-                ' has been kicked from following channels : ' . implode(',', array_diff($channels, $allowedChannels));
-            $slackLog->save();
+            if (!empty($extraChannels)) {
+                $this->processChannelsKick($slackUser, $extraChannels);
+                $this->logEvent('kick', $extraChannels);
+            }
 
             // remove granted channels from channels in which user is already in and kick him
-            $this->processGroupsKick($slackUser, array_diff($groups, $allowedGroups));
-
-            $slackLog = new SlackLog();
-            $slackLog->event = 'kick';
-            $slackLog->message = 'The user ' . $this->user->name .
-                ' has been kicked from following channels : ' . implode(',', array_diff($groups, $allowedGroups));
-            $slackLog->save();
+            $allowedGroups = $this->allowedChannels($slackUser, true);
+            $extraGroups = array_diff($groups, $allowedGroups);
+            if (!empty($extraGroups)) {
+                $this->processGroupsKick($slackUser, array_diff($groups, $extraGroups));
+                $this->logEvent('kick', $extraGroups);
+            }
         }
 
         return;