+++ src/Models/SlackChannel.php
@@ -7,7 +7,7 @@ use Illuminate\Database\Eloquent\Model;
 class SlackChannel extends Model
 {
     protected $fillable = [
-        'id', 'name', 'is_group'
+        'id', 'name', 'is_group', 'is_general'
     ];
 
     protected $primaryKey = 'id';